add test class
