+++ weather/README.md
@@ -12,3 +12,18 @@ This repository contains the source code for the [Building Modern Web Applicatio
 
 ## Text tutorial
 You can find a text version of the tutorial in the [Vaadin Documentation](https://vaadin.com/docs/latest/flow/tutorials/in-depth-course).
+
+
+working area
+user login
+username:user
+pwd: demoUser
+admin login
+username: admin
+pwd: demoAdmin
+
+-----
+add flyway for data migration
+add lombok
+add postgres sql 
+add vaadin framework
