+++ weather/src/test/java/com/eastnetic/service/AppApplicationTests.java
@@ -1,0 +1,13 @@
+package com.eastnetic.service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class AppApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
